requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.1
SQLAlchemy>=2.0.35
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.11
supabase>=2.22.1

# AI Features (optional - only needed if enabling AI features)
anthropic>=0.18.0
